move user db path utils to UserDBPath
and delete ApplicationUtils and TestConnectionHandler
