+++ src/test/java/tests/CaseTests.java
@@ -12,7 +12,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import tests.sandbox.TestConnectionHandler;
 import util.Constants;
 import utils.TestContext;
