+++ src/test/java/tests/FilterTests.java
@@ -1,20 +1,18 @@
 package tests;
 
 import application.SQLiteProperties;
-import beans.*;
-import sandbox.SqlSandboxUtils;
-import sandbox.SqliteIndexedStorageUtility;
-import sandbox.UserSqlSandbox;
+import beans.SyncDbResponseBean;
 import org.commcare.cases.model.Case;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import tests.sandbox.TestConnectionHandler;
+import sandbox.SqlSandboxUtils;
+import sandbox.SqliteIndexedStorageUtility;
+import sandbox.UserSqlSandbox;
+import sqlitedb.UserDB;
 import util.Constants;
 import utils.TestContext;
-
-import static org.mockito.Matchers.any;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = TestContext.class)
@@ -30,7 +28,7 @@ public class FilterTests extends BaseTestClass {
         assert(syncDbResponseBean.getStatus().equals(Constants.ANSWER_RESPONSE_STATUS_POSITIVE));
         assert(SqlSandboxUtils.databaseFolderExists(SQLiteProperties.getDataDir()));
 
-        UserSqlSandbox sandbox = new UserSqlSandbox(new TestConnectionHandler(SQLiteProperties.getDataDir() + "synctestdomain/synctestuser"));
+        UserSqlSandbox sandbox = new UserSqlSandbox(new UserDB("synctestdomain","synctestuser", null));
 
         SqliteIndexedStorageUtility<Case> caseStorage =  sandbox.getCaseStorage();
